requirements.txt
# FastAPI and ASGI server
fastapi==0.115.6
uvicorn==0.34.0

# HTTP client for external API calls
httpx==0.28.1

# Authentication and security
bcrypt==4.3.0
PyJWT==2.10.1

# Data validation and serialization
pydantic==2.10.4
pydantic-settings==2.8.1

# Environment management
python-dotenv==1.0.0
